Added UART communication for debugging. Fixed uint8_t casting error resulting in erroneous duty cycle calculations. Verified with scope PWM duty cycle and direction control.

+++ haptic/spi_test.h
@@ -26,5 +26,5 @@ uint16_t pwm_duty = 0;      // Safely start with no commanded motor motion
 unsigned char pwm_direction = 1;  // Initialize motor for forward motion
 _PIN* PWM_I1 = &D[8];       // Input 1 to motor driver chip
 _PIN* PWM_I2 = &D[7];       // Input 2 to motor driver chip
-uint32_t DUTY_MAX = 65536;  // Value used for converting between uint16_t
+long DUTY_MAX = 65536;  // Value used for converting between uint16_t
                             // fractional representations and float percentages.
